+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 ### Fixed
 
 - long chapter names alignment in history sheet
+- full player landscape mode
 
 ## [v0.8.1] - 2026-09-08